+++ tests/integration/specs/integrate/claude.test.ts
@@ -25,7 +25,7 @@ import { isAbsolute } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
 
-import { SQAA_INSTRUCTIONS_SUBFEATURE_ID } from '@/commands/integrate/_common/sqaa-entitlement.ts';
+import { SQAA_INSTRUCTIONS_SUBFEATURE_ID } from '@/commands/integrate/_common/features/sqaa-instructions-feature.ts';
 import { VORTEX_FEATURE_ID } from '@/commands/integrate/_common/vortex.ts';
 import { claudeIntegration } from '@/commands/integrate/claude/declaration.ts';
 import { buildLocalBinaryName } from '@/core/host/install/secrets.ts';
